Fix CI build errors against Flutter 3.47 / image 4.9.2 / printing 5.15
- Scanner now decodes/resizes via dart:ui only (instantiateImageCodec +
  drawImageRect + toByteData raw RGBA) — removes image-package
  copyToFormat/FORMAT_L8/ByteOrder usage that no longer exists in 4.9.2
- ui.Paint() is non-const in this Flutter version: drop const
- Overlay encodes to PNG (proven ImageByteFormat member) instead of jpg
- PDF view uses the static Printing.layoutPdf preview (no instance ctor)
- copyRotate angle is a named argument in image 4.9.2
- Fix _dirFor return type (Directory), wrap fallback corner in
  DetectedCorner, remove all num->double/int mixes (clamp results,
  switch-expr scores, otsu bestT)

+++ lib/services/omr/omr_geometry.dart
@@ -19,9 +19,9 @@ import 'dart:ui' show Offset;
 class OMrGeometry {
   OMrGeometry(this.total)
       : assert(total >= 1 && total <= 100, 'OMR supports 1–100 questions') {
-    _columns = ((total + maxPerColumn - 1) ~/ maxPerColumn).clamp(1, 4);
+    _columns = ((total + maxPerColumn - 1) ~/ maxPerColumn).clamp(1, 4).toInt();
     perColumn =
-        ((total + _columns - 1) ~/ _columns).clamp(1, maxPerColumn);
+        ((total + _columns - 1) ~/ _columns).clamp(1, maxPerColumn).toInt();
     questionGap = 8.0 * k;
     final rawW = (contentW - (_columns - 1) * questionGap) / _columns;
     questionWidth = rawW > naturalRowW ? naturalRowW : rawW;
